Extract EC disaggregation helpers into a mixin
Move `_make_producer_output`, `_run_encoder_and_save`,
`_run_decoder_with_cached_encoder`, and `_get_mm_hash_for_request` out
of `RBLNOptimumModelRunner` and into a new
`vllm_rbln/v1/worker/ec_disagg_helpers.py` as `ECDisaggHelpersMixin`.
All call sites remain `self.<method>(...)` — behaviour is unchanged;
only the module/class declaration layout moves.
